Extract to ts
