+++ app/gui/src/project-view/components/SelectionMenu.vue
@@ -22,10 +22,10 @@ const pickColorMulti = resolveAction('components.pickColorMulti')
       }"
     />
     <ActionButton action="components.alignLeft" />
+    <ActionButton action="components.alignCenter" />
     <ActionButton action="components.alignRight" />
     <ActionButton action="components.alignTop" />
     <ActionButton action="components.alignBottom" />
-    <ActionButton action="components.alignCenter" />
     <ActionButton action="components.copy" />
     <ActionButton action="components.deleteSelected" />
     <ColorPickerMenu
